commit PDM of Industrial Equipment Using ML & DL
The goal of the "Predictive Maintenance of Industrial Equipment (PDM)" project is to develop a PDM 
model to optimize the maintenance process for industrial equipment. To forecast when equipment is 
likely to fail in the future, the model will analyze historical data on equipment failures and maintenance 
actions., the project will make use of statistical analytic methods and Machine learning (ML) 
algorithms and Deep Learning Neural Networks.